Remove handicap calculator, migrate stats calculations to core
