+++ src/main/java/dev/tk2575/golfstats/core/stats/ApproachCategory.java
@@ -1,4 +1,4 @@
-package dev.tk2575.golfstats.details.api.stats;
+package dev.tk2575.golfstats.core.stats;
 
 import dev.tk2575.golfstats.core.golfround.shotbyshot.Lie;
 import dev.tk2575.golfstats.core.golfround.shotbyshot.Shot;
