+++ src/main/java/dev/tk2575/golfstats/core/stats/SimpleStat.java
@@ -1,4 +1,4 @@
-package dev.tk2575.golfstats.details.api.stats;
+package dev.tk2575.golfstats.core.stats;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
