--- src/main/java/dev/tk2575/golfstats/details/api/handicapcalculator/TeeDTO.java
@@ -1,14 +1,0 @@
-package dev.tk2575.golfstats.details.api.handicapcalculator;
-
-import lombok.*;
-
-import java.math.BigDecimal;
-
-@RequiredArgsConstructor
-@Getter
-public class TeeDTO {
-	private final String name;
-	private final BigDecimal rating;
-	private final BigDecimal slope;
-	private final Integer par;
-}
