+++ src/main/java/dev/tk2575/golfstats/details/api/stats/ShotAnalysis.java
@@ -1,6 +1,7 @@
 package dev.tk2575.golfstats.details.api.stats;
 
 import dev.tk2575.golfstats.core.golfround.shotbyshot.Shot;
+import dev.tk2575.golfstats.core.stats.StatsApiValueSupplier;
 import lombok.*;
 
 import java.math.BigDecimal;
